Add VtePty and adapt the VteTerminal APIs to it
Add VtePty as a GObject holding the info about the PTY. Add new API to
VteTerminal to set a VtePty, and deprecate the old API that takes a FD
to the PTY. Also deprecate the whole of the undocumented _vte_pty_*()
APIs.

Add vte_terminal_fork_command_full() variant that allow providing a
custom child setup function and that returns a GError on failure.

Bug #585841, bug #320128, bug #514447, bug #588871.