Use accessors for gtk objects
This change introduces vte-gtk-compat.h to allow compatibility with the
current GTK+ requirement (2.14+) while allowing us to use GSEAL when
available.  When the GTK+ requirements are moved up to 2.20+, it should
be possible to entirely remove the header.

Setting of the GTK adjustment fields still occurs without the accessors
as setting the value and then using value_changed may introduce
different behaviour and/or double signal emissions.