Report correct minimum/natural sizes for GTK+ 3
In GTK+ 3, the hack of reporting the current size of the terminal
as its minimum size breaks down because GTK+ 3 will never allocate
a widget smaller than its minimum size. However, in GTK+ 3 it's
possible to report separate minimum and natural sizes, so we
should just report a minimum size of 1x1 character and a natural
size of the current size.

This requires a change in vteapp.c to make the window start off
at the right size.

https://bugzilla.gnome.org/show_bug.cgi?id=631903