BGO#654630: Fix accessibility with gtk+ 3.1
The gtk+ accessibility code was reworked in the 3.1 cycle, and the code
to fetch the accessibility type associated with GtkWidget no longer
works.  At present, third-party ATK implementations for GTK widgets are
expected to derive from GtkAccessible, which implements none of the
accessibility code for standard gtk+ widgets, so copied/adapted much of
this code from the GtkWidgetAccessible implementation.