Move definition of VteCharAttributes up, #676999
I found the element-type for the attributes in vte_terminal_get_text
vte_terminal_get_text_include_trailing_spaces and vte_terminal_get_text_range
strange to be of type "Vte.CharAttributes" and not VteCharAttributes but when I
changed it I only got a warning. Then I saw that there was already a warning
about "Unknown namespace for identifier 'vte_char_attributes'", so I moved the
definition of VteCharAttributes and now those functions are introspectable.