Use @basename@ rather than @filename@ in vtetypebuiltins.h
On Debian/Ubuntu-based systems it is helpful for headers to be identical
across architectures, thereby allowing the -dev packages from multiple
architectures to be installed simultaneously; in particular, this
simplifies cross-building.  When building out-of-tree, the use of
@filename@ causes the build directory to end up in the generated header
file, which means that it isn't reliably identical on all architectures.
Using @basename@ instead fixes this problem.

Although it isn't necessary to change the corresponding .c file as well,
I did so for consistency.

https://bugzilla.gnome.org/show_bug.cgi?id=690554