emulation: Fix the scrollbar position with restricted scrolling
To reproduce, produce at least a screenful of text so that the scrollbar is
active. Then enable restricting scrolling, with the upper bound being the
topmost row, e.g. echo -e '\e[1;15r'. Now produce more output so that the
cursor hits the boundary at the 15th row and the content starts scrolling.
The scrollbar starts wandering upwards, though it should stay at the bottom.
When trying to scroll by Shift+PageUp or by dragging the scrollbar, suddenly
everything jumps back to normal.

https://bugzilla.gnome.org/show_bug.cgi?id=710488