emulation: Fix background when bottom row is scrolled upwards
Make scrolling the bottom row upwards not fill this row with the
background color. Only the new empty bottom line should be filled.

To reproduce, run

$ echo -e '\e['$LINES'H\e[43mBottom row'; sleep 100

When the row containig the text "Bottom row" gets scrolled up by 1, it was
filled with the background color in vte, but not in xterm.

https://bugzilla.gnome.org/show_bug.cgi?id=710486