ring,stream: Refactor the interface between the ring and the stream
Refactor and simplify the interface between ring and vte-stream; this is
required for the forthcoming rewrapping feature. Paging is now a concept
private to vte-stream. Instead of telling when to turn page, the ring now
advances the tail of the streams, and vte-stream does paging on its own
based on this information.

https://bugzilla.gnome.org/show_bug.cgi?id=336238#c74