docs: Fix licence header
The library is LGPL, and the docs include the docs comments from the library
sources, so it's logical that the docs licence should be LGPL too. The docs
already said so in the text, but the licence header in the comments said GFDL.
Since I'm the only one who's contributed to the xml file itself, I just
changed this to LGPL, too.

https://bugzilla.gnome.org/show_bug.cgi?id=728795
(cherry picked from commit 957a85b11750362a265526d63ca1423ae453cf3b)