build: Work around include directory issue
The correct way to include vte.h is <vte/vte.h> but the source directoy
is named src/ not vte/ so this doesn't work while building vte itself,
and it breaks when building the vala test application.

Make vte/vte.h symlink to src/vte.h, and dist that. Ugly, but it works™.