emulation: Implement DECSET 1004
Add support for xterm's DECSET 1004 to enable focus tracking.
This sends CSI I on focus in, and CSI O on focus out.

Contrary to xterm, vte disables the focus tracking on reset, and when
focus tracking is enabled, immediately sends a focus in or focus out, so
that clients always have an accurate value.