regex: Pass more data to regex matching
Pass the whole match contents before the line to matching, which is useful
if there are lookbehind assertions in the regex. We also want to pass
the whole contents after the line to matching to satisfy lookahead
assertions, but this isn't possible currently without new PCRE2 API.

This also fixes the cell row/column info in the debug output.