lib: Add grid coordinate and span types
Add vte::grid:coords to hold a (column, row) pair, and vte::grid::span
to hold a pair of (start, end) coordinates.

They will be used to replace passing separate (column, row) pairs around
in arguments and data structures.

Also contains tests for the member functions, and static assertion
to test the size of these structs, and ensure they are and stay POD.