widget: Change last-line newline behaviour of get_text()
To get all the text on the last line to the right, you can either
pass m_column_count (or MAXLONG), *or* add 1 to the end row, and
pass -1 as last column. Make it so that the difference between these
is that the former does *not* append a final newline character,
while the latter does append a final newline character.

This fixes a frequent assert in vte_terminal_accessible_text_scrolled(),
but the assertion in that function will need to be fixed, too, since
it appears the assert can be triggered in other ways too.

https://bugzilla.gnome.org/show_bug.cgi?id=759319