widget: Don't deselect when updating the clipboard
gtk_clipboard_set_with_data() differs from gtk_clipboard_set_with_owner()
in that the old clear callback is *always* invoked instead of only
when the owner (user_data) is different from the last owner (user_data).
I think this is a gtk+ bug, but that code has been that way since 2000,
so who knows why it does it this way.

Adapt to this by guarding the clear callback to gtk_clipboard_set_with_owner()
with a bool that is set/unset around the call and makes the callback
bail out early if we're updating the clipboard from within vte.

https://bugzilla.gnome.org/show_bug.cgi?id=762446