emulation: Swallow urxvt OSC 777
Accept OSC 777 but make it a no-op.

Some distros patch their vte downstream with the rejected patch from
bug 711059, which leads to problems with these escape sequences being
shown on the terminal when using an unpatched upstream build of vte
but the system-installed patched vte.sh script, due to vte not
ignoring unknown escape sequences (bug 403130).

https://bugzilla.gnome.org/show_bug.cgi?id=403130
https://bugzilla.gnome.org/show_bug.cgi?id=711059