lib: Add async spawning
Add vte_pty_spawn_async/finish for async spawning of a
child process on a VtePty, and a vte_terminal_spawn_async
convenience function that wraps creating the PTY and spawning
the child process, and can cope with the terminal being
destroyed between starting the spawning and the operation
being finished.

This is a partial solution to bug 772354; a full solution
will require glib changes.

https://bugzilla.gnome.org/show_bug.cgi?id=772354