Fix the generation of the signal marshallers
The generation of the signal marshallers is based on an undefined
behaviour of the glib-genmarshal tool. Passing both the --header and
--body command line arguments is undocumented and untested behaviour.
The generated C source adds prototypes for the marshallers, and wraps
the whole file between C++ guards.

Instead of relying on undocumented behaviour, we can achieve the same
result by including the generated header in the generated source.

https://bugzilla.gnome.org/show_bug.cgi?id=784853