emulation: Simplify sequence handlers and parameter handling
Encapsulate the use of the deprecated GValueArray API in a
helper class, and use that to extract the parameters passed to
the sequence handlers.

Also, move from using 2-stage sequence handlers that first call a
static function which in turn calls a method on VteTerminalPrivate
to using the member functions directly via a member function pointer.