all: Rearrange VteCellAttr to store all colours together in one uint64_t
This re-arranges the VteCellAttr struct to store the foreground,
background and decoration colour together in one uint64_t. This will
allow later optimising the *other* attributes.

A 24-bit (8 bit per component) color is stored in 25 bits in vte
(one extra bit is necessary to distinguish direct colours from
index colors), so to fit 3 colours into 64 bits, it was necessary
to reduce the depth of the (lesser used) decoration colour to
4 bit per component.