emulation: Clamp cursor column
When setting/unsetting tab stops, make sure to clamp the cursor
column to 0..m_column_count-1. This is necessary because vte's
weird autowrap handling where it puts the cursor on column
m_column_count if there's a wrap pending, instead of keeping
the cursor at the right margin and using a flag, like every other
terminal emulator does.

Fixes a crash when setting/unsetting a tabstop while the cursor
is in that pending autowrap state.