widget: Always emit the child-exited signal
When there is still a child process running when the
VteTerminal is destroyed, it gets kill()ed, but since
the child-exited signal was only emitted from the
VteReaper callback, it was never actually emitted, since
by that time VteTerminal has ceased to exist.

Instead of waiting for the reaper, immediately emit
the child-exited signal in VteTerminal::dispose(). The
exit status is synthesised to be WIFSIGNALED() with
WTERMSIG() == SIGKILL.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/16