draw: Optimise invalidation and painting
Cells may contain content that draws outside the cell area. Limit this
to only allow drawing up to the underline position of the line above,
and the overline position of the line below, by clipping the cairo
context to that area before drawing each line.

Only paint lines that are in the invalidated region (or which due to
overdrawing may draw into the invalidated region).

https://gitlab.gnome.org/GNOME/vte/issues/56