lib: Rework child exit and EOF handling
When the child process exits, we used to immediately unset the PTY,
which causes us to miss data written by the child but not yet read
by vte.

Instead, only store the child exit status, and defer emitting the
'child-exited' signal until after all the pending data has been read
and processed.

Similarly, rework how EOF is processed. Instead of immediately
queuing the emission of the 'eof' signal, only take note of the EOF,
and process it after all pending data has processed. There also was
a bug in that we took the first occurence of G_IO_HUP in
Terminal::pty_io_read() to stop reading more data. Instead, only
take a pure G_IO_HUP without G_IO_IN as EOF, or if reading data
from the PTY returns the EIO error.

This also fixes the bug where a(ny) partial character(s) not yet fully
decoded by the UTF-8 and ICU decoder would not show in the output; this
now correctly flushes the decoder, which inserts either a replacement
character (for the UTF-8 decoder) or the character(s) in the ICU decoder
internal state (most likely also a replacement character).

https://bugzilla.gnome.org/show_bug.cgi?id=777686