lib: Use ICU for legacy charset support
Instead of converting the whole chunk of input from the input
charset to UTF-8 in one go, we need a decoder that consumes the
input one byte at a time. Since the iconv API is not particularly
suited to this (or, really, any) task, switch to using ICU for this.

Add functions to get the list of supported legacy charsets, and
to check whether a particular string is a supported charset.

Fixes https://gitlab.gnome.org/GNOME/vte/issues/40