pty: Fix indefinite wait for EOS after child-exited
When the child process exits, we wait for EOF on the PTY, so that we
can read and process any data already emitted by the child process but
not read yet at that time.

However that leads to a problem when the child process created child
processes of its own that inherited the PTY as stdin/out/err, which keeps
the PTY open and thus we get no EOF until all such processes have exited.

To work around this, install a timeout when the child process exits, and
when the timer times out, fake en EOF.

https://gitlab.gnome.org/GNOME/vte/issues/204