icu: glue: Set error callback on the from-unicode conversion
The from-unicode side of the converter is only used on user input
(keyboard, IM, clipboard paste). When it receives input not representable
in the target encoding, it inserts the substitution character of the
target encoding, which in many cases is U+001A. However, when sent to a
PTY, that is interpreted as Ctrl-Z.

Instead, set the from-unicode converter to error out when encountering
such unrepresentable input.

Fixes: https://gitlab.gnome.org/GNOME/vte/issues/233