widget: Define text_view_bg in fallback CSS if the theme doesn't
Adwaita defines text_view_bg since GTK 3.24.22, but non-Adwaita themes
such as the built-in HighContrast[Inverse] and the third-party Numix
don't necessarily. Unfortunately, if it isn't defined, GTK treats that
as equivalent to it being fully transparent, leading to misrendering
that often makes terminal text unreadable.

Use a fallback style provider to make sure text_view_bg is always
defined, using the theme base colour if necessary. This is done at a
low priority, so that themes like Adwaita have the opportunity to
override it.

Resolves: https://gitlab.gnome.org/GNOME/vte/-/issues/284
Signed-off-by: Simon McVittie <smcv@debian.org>