widget: Add API to disable fallback scrolling
Ordinarily, when the VteTerminal hasn't consumed a scroll event otherwise
(e.g. by sending the event to the application), it falls back to perform
a history scroll.  When the terminal is added to a GtkScrolledWindow, this
result in the scrolled window not performing kinetic scrolling.

This commit adds API (and GObject property) to VteTerminal to disable the
fallback scrolling.

Based on a patch by Tony Houghton <h@realh.co.uk>.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/234