font: pangocairo: Remove redundant code
The max_height variable is initialized as 1, and then iteratively
updated to track the maximum height of all the characters from U+0021
to U+007E. The final value is assined to m_height. Hence m_height can
never be zero.

In the case of m_ascent, the contents of m_layout don't change between
the two places where it's assigned. Therefore, the second assignment is
redundant.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/355