fonts: More metrics sanity checks
The pango font metrics for some fonts return a height lower than the
height measured from the U+0020..U+007E characters. Using the metrics
in this case may lead to the descenders getting cut off from characters
in the last line, depending on the padding available.

Change the code to only use the pango metrics when its height is at
least that of the measured characters.

Fixes: https://gitlab.gnome.org/GNOME/gnome-terminal/-/issues/340