widget: Fix padding borders
On gtk3, the padding needs to be added to the widget's size request, and
is inside its allocation, while on gtk4, the padding is outside the
widget's allocation.  The inner padding (use to [xy]align and [xy]fill
properties) is always inside the allocation.

Also fixes event coordinate translation on gtk4 to take the inner border
into account.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2554