widget: gtk4: Hint and quantize font metrics
With gtk4, the font metrics are unhinted by default. This causes
problems when measuring the font, since the width or height may
have fractional pixel values that get rounded up - resulting in
vte using integer cell sizes that are larger than would be
appropriate for the font - the usual side-effect is that horizontal
letter spacing looks too wide.

To fix this, set the cairo font option to hint vertical metrics,
which will result in the glyph ascent and descent values both being
integers.

Also make pango round glyph positions, so that the horizontal advance
of the glyphs will be integers appropirate for a renderer that doesn't
do sub-pixel glyph positioning.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2573