fonts: keep layout text available for cairo
When a cairo recording surface is used, it will memcpy() the text and
therefore expects the layout's text to be complete enough to contain each
glyph item's offset from base pointer.

This keeps that text around long enough to satisfy that requirement. It
will be reset on the next operation/call to get_unistr_info().

Fixes #2620