fonts: Keep the PangoLayout unchanged
The previous commit still wasn't completely fixing the problem, since
the FontInfo and the UnistrInfo are cached too, and m_layout will be
re-used when creating the next UnistrInfo, but the code in
pango_cairo_show_layout_line() requires the PangoLayout to still have
the same text as when it was created for the UnistrInfo.

To fix this, adopt the PangoLayout into the PangoLayoutLine cached in
the UnistrInfo, and create a new layout for next use.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2606
Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2620
(cherry picked from commit 3c8f66be867aca6656e4109ce880b6ea7431b895)

Fixes: https://gitlab.gnome.org/Teams/Releng/freeze-breaks/-/issues/131