widget: a11y: Optimize text changes on scrolling with modifications
In 036d337cb84c ("widget: a11y: Add missing text changes on scrolling
with modifications") we fixed text updates for the scroll+modification
case by updating the whole terminal. Updating the whole terminal however
makes Orca very talkative.

This adds support for simple diff support for the common case: an ample head
and/or tail content is the same. This allows to emit only the modification,
which is usually relatively small.

Fixes #88