unichar: hoist common unichar width into caller
The most common case we run into here is the < 0x80 which will always be
a cell width of 1. The compiler does not seem to be inlining even
portions of _vte_unichar_width() naturally, so force it with a macro to
hoist just the common hot path into the caller.

This is good for about 1% of savings similar to the previous inlines.