unistr: inline unichar into GString as UTF-8
This code path is *extremely* hot. This patch brings over the UTF-8
conversion and uses g_string_append_len() which is now inlined by GLib
to allow for simply memcpy()'s the utf8 bytes into the destination string.

The cost of calling g_string_append_unichar() is extremely high. It
currently is treated the same as a g_string_insert_unichar() at the end
which does not allow for the same level of optimization as a dedicated
code path for append().

Additionally, UTF-8 conversion does some of the same checks which the
calling code would have already done. Some of that may be able to be
shared when inlined into the caller.

Even with other string related patches applied, this is good for a couple
more percent of CPU cost savings.