ring: inline vte::base::Ring::index_writable()
This function gets called a lot from various places, and previously was
done through a C wrapper. It is a good candidate for inling, but cannot
do that while the C wrapper is in use.

So both make it inlined and remove the C wrapper since it was likely just
there for transition purposes anyway.