ring: remove VteRing C wrapper for vte::base::Ring
Part of transitioning to vte::base::Ring was a intermediate _vte_ring_*()
wrapper API. This C API was preventing the compiler from inlining various
API that was desired to be inlined.

This removes the C wrapper and updates all of the callers to use the
appropriate vte::base::Ring() methods instead.

Additionally, functions which were intended to be inlined have been
inlined and do indeed remove some of the Ring overhead in profiles.