parser: move vte_parser_t into vte::parser::Parser
This is just the first step towards being able to inline the very hot
paths from vte.cc through the Parser structure. Currently everything is
inlined but my guess is that we'll want to be selective about that and
make some things not inline and other things hoisted through an inline
function.

Performance is roughly the same as before, except that I notice ever so
slightly more frames are rendered with this version when using the GTK 3
implementation.