build: compile out assertions and cast-checks
In production builds, many GNOME libraries compile out both assertions and
cast checks. These often have significant overhead in tight loops which
we don't need in production releases but are critical during the
development phase to avoid shipping obvious bugs.

This does the same for libvte libraries.

Production builds are anything not starting with 'debug' in the profile.
That includes "plain" as well as "release" as distributions almost always
use "plain" with their custom compiler flags.

Testing with an extremely large file cat to vte-2.91 this yields an 8%
improvement compared to the previous commit.