vteinternal: use specialized array for char attributes
When the mouse pointer is within the widget we do a lot of of querying for
the text (each changing frame) due to match_hilite_update(). This spends
a few percent simply doing GArray sizing calculations.

This patch does like we've done elsewhere and saves a bit over 3% CPU when
running btop at 10hz w/ the mouse pointer within the widget.