terminal: require caller to allocate GString for get_text()
This pushes the allocation of the GString for Terminal::get_text() and
associated functions to the caller. In doing so, this allows reuse of
m_match_content. Since a maximized window can easily contain a page or
more of data, avoiding the power-of-two growth and GString maybe_expand()
checks here is useful. It combines well with the previous commit which
allows for reuse of the m_match_attributes array for the same reason.

You'll mostly see this behavior when the mouse is sitting over the
terminal widget and there are a lot of updates to the screen.