draw: implement native GTK 4 drawing
This abstracts DrawingContext into an abstract base class with two
implementations: DrawingCairo and DrawingGsk. The DrawingCairo version
is only compiled on GTK 3 and keeps the design largely the same as it was
before, using one large Cairo context. The DrawingGsk attempts to provide
the same feature set but uses native GtkSnapshot and GskRenderNode
features.

Minifont has been changed to be more performance friendly in GTK 4. Cairo
surfaces are maintained in a sized cache based on size and character. They
are drawn onto the canvas with as a mask for color which is slightly
different on GTK 3 to how they were drawn before. On GTK 4 they are also
drawn as a mask for color. A new fast path has been added to the GTK 4
renderer to improve performance with this pattern. It, fundamentally, is
the same pattern GTK uses internally for it's glyph cache.

For GTK 4, only the PangoGlyphString coverage is used. This allows the
creation of GskTextNode directly from the built strings.

There are possibilities to improve performance further by caching the
container GskRenderNode for each row and re-using them frame to frame.
However, GTK will determine the appropriate damage region for the frame
regardless.

A new Rectangle struct has been added to allow calling code to create the
rectangle in the best format used by the underlying rendering technology
without further conversions.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2553