emulation: Adjust cursor tracking and autowrapping for the right margin
With right margin support, we can no longer rely on the cursor being off
to the right to denote that the next character will wrap. An additional
boolean is required to distinguish this from the cursor being explicitly
moved to just beyond the right margin.

Fixes https://gitlab.gnome.org/GNOME/vte/-/issues/2677