widget: gtk4: Delay unsetting the context menu to idle
The closed signal on the popover comes before the action callback for
the chosen action is dispatched, and unsetting the menu makes the
action fail to resolve.  Hack around this design flaw by delaying the
unsetting to an idle handler.

Remove the notify::visible handler since it runs just after the closed
signal anyway.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2716