widget: a11y: Add missing text changes on scrolling with modifications
When vte_terminal_accessible_text_scrolled gets called, the terminal
might have not only scrolled, but also changed.  We thus have to check
whether the remaining text really is exactly the same as expected.

We here support simple diff for the common case: an ample head and/or tail
content is the same. This allows to emit only the modification, which is
usually relatively small.

Fixes: https://gitlab.gnome.org/GNOME/vte/issues/88