widget: notify of style change upon css_changed
The org.gnome.desktop.interface text-scaling-factor setting used by
the "Large Text" option in control-center applies to the gtk-xft-dpi which
in turn effects the CSS machinery.

We must notify of a style change so that the PangoContext used by our
PangoLayout to measure and cache PangoFont will take the new scaling
factor into account.

This makes GTK 4 act the same as GTK 3 did (and still does) when the
setting changes.