fonts: use CAIRO_HINT_STYLE_NONE on GTK 4
This fixes an issue when fractional scaling is used that a portion of a
hinted glyph could fall outside the ink rect due to pixel grid shifting.
This is more apparent with the Vulkan/NGL GPU renderers in GTK now that
they are aware of the fractional scaling API of Wayland. Without this
you risk a small amount of unintended clipping. See GNOME/gtk#6367.

See the following link for additional details. However, in the case here
with fractional scaling, disabling hint-metrics is not an option because
it will result in the same output as hint-metrics: on with
hint-style: slight.

https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6823#note_1989289

Fixes #2736