emulation: Allow default parameter value for XTVERSION
There is a discrepance between the xterm documentation and the actual
xterm behaviour in that the documentation does not mention that the
sequence admits a default value of 0 (to report the xterm name and
version), while the code does allos the user of a default param.

Make vte follow the xterm code.

https://gitlab.gnome.org/GNOME/vte/-/issues/235