terminal: inline hot path of cell_is_selected_log()
This function appears to be called a lot, especially during high scrolling
scenarios (such as ctrl+f/ctrl+b in vim). There is a typical fast path
which is that ring is updated and there is no selection. Hoist that into
the callers and leave the selection case behind a function call.

This shaves about 1.5% of application samples off time spent in
Terminal::draw().