emulation: DECCRA fixes
Use basic_cell to fill before the dest column.

When copying from a source row that ends in a double-width character that
doesn't fit fully into the source rect, still copy the attributes to the
destination but with erased content.

When copying from a source row that ends before the right side of the
rect, copy default attributes to the dest.