drawing: gsk: Draw cell background using scaled texture
This abstracts the drawing of cell backgrounds so that the GTK 4 variant
of DrawingContext may use a scaled texture of 32 bit-per-pixel where the
number of pixels is equivalent to the number of cells.

By drawing the background this way, we allow OpenGL/Vulkan to scale the
texture with nearest-neighbor to avoid seams between adjacent rectangles.

This fixes an issue caused by the introduction of fractional-scaling
no longer drawing at 2x and downscaling in GTK 4.14.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2778
Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2773
Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2825