vteaccess-gtk4: Return a range per text attribute
The docs say that a text attribute is composed of a range,, name and value.
We were returning only one range, though, so the code in Gtk counting on
the fact that n_ranges was also the number of attributes, was confused.
As a result, we were always only returning the first text attribute.