build: Add option to not build the vte application
This allows distributions to build vte with and without the app
in order to package the app separately.

Note that the vte application is not simply a demo but a fully
functional terminal application in its own right, and also required to
debug issues before filing any issue by end users or distribution
packagers. As such, distributions are encouraged to provide a package
with the vte application.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/merge_requests/5